Closes-Bug: #1471596.             Supports HA monitoring even in cases where             keepalived is not configured to provide VIP
            Patch-1: Upgrade support added
            Patch-2: Added cmon conf changes

+++ contrail_provisioning/common/scripts/contrail-rmq-monitor.sh
@@ -9,6 +9,8 @@ readonly PROGNAME=$(basename "$0")
 readonly LOCKFILE_DIR=/tmp/ha-chk
 readonly LOCK_FD=200
 
+readonly command=$1
+
 LOGFILE=/var/log/contrail/ha/rmq-monitor.log
 RMQ_CLUSTER_OK="running_nodes"
 RMQ_RESET="service rabbitmq-server restart"
@@ -21,8 +23,25 @@ rstcnt="/tmp/ha-chk/rmq-rst-cnt"
 numrst="/tmp/ha-chk/rmq-num-rst"
 rmqstop="supervisorctl -s unix:///tmp/supervisord_support_service.sock stop rabbitmq-server"
 killbeam="pkill -9  beam"
-killepmd="pkill -9 epmd"
+killepmd="pkll -9 epmd"
 rmmnesia="rm -rf /var/lib/rabbitmq/mnesia"
+cleanuppending="/tmp/ha-chk/rmq_mnesia_cleanup_pending"
+sethapolicy="rabbitmqctl set_policy HA-all \"\" {\"ha-mode\":\"all\",\"ha-sync-mode\":\"automatic\"}"
+STOP="STOP"
+MYIPS=$(ip a s|sed -ne '/127.0.0.1/!{s/^[ \t]*inet[ \t]*\([0-9.]\+\)\/.*$/\1/p}')
+MYIP=0
+MONITOR="MONITOR"
+RMQ_CLNTS=${#RMQ_CLIENTS[@]}
+
+for y in $MYIPS
+ do
+  for (( i=0; i<${DIPS_SIZE}; i++ ))
+   do
+     if [ $y == ${DIPS[i]} ]; then
+        MYIP=$y
+     fi
+   done
+done
 
 if [ ! -f "$LOCKFILE_DIR" ] ; then
         mkdir -p $LOCKFILE_DIR 
@@ -40,16 +59,16 @@ if [ ! -f "$rstinprog" ] ; then
          touch "$rstinprog"
 fi
 
+if [ ! -f "$cluspart" ] ; then
+          touch $cluspart
+fi
+
 if [ ! -f "$rstcnt" ] ; then
-          touch $rstcnt
+        touch $rstcnt
 fi
 
 if [ ! -f "$numrst" ] ; then
-          touch $numrst
-fi
-
-if [ ! -f "$cluspart" ] ; then
-          touch $cluspart
+        touch $numrst
 fi
 
 timestamp() {
@@ -164,7 +183,7 @@ do
  (rabbitmqctl list_channels | grep $hosts | wc -l > "$file") & pid=$!
  (rabbitmqctl cluster_status | grep -A 1 running_nodes > "$cluschk") & pid1=$!
  (rabbitmqctl cluster_status | grep partitions | grep ctrl | wc -l > "$cluspart") & pid2=$!
- log_info_msg "pidof pending rmq channel $pid and cluster check $pid1"
+ log_info_msg "Checking for stable state of rmq channel (monitored by pid $pid) and cluster (monitored by pid $pid1)"
  sleep 10
  if [ -d "/proc/${pid}" ]; then
   (exec pkill -TERM -P $pid)&
@@ -177,6 +196,77 @@ do
  fi
   i=$[$i+1]
 done
+}
+
+cleanup()
+{
+ dst=$1
+ if [[ $dst != '' ]]; then
+   out=$(ssh -o StrictHostKeyChecking=no -o ConnectTimeout=5 $dst "date")
+   if [[ $out != '' ]]; then
+     (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=15 $dst "$rmqstop")
+     (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=15 $dst "pkill -9 beam")
+     (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=15 $dst "pkill -9 epmd")
+     (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=15 $dst "$rmmnesia")
+     (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=15 $dst "$RMQ_RESET")
+     log_info_msg "Cleaned up mnesia and reset RMQ on $dst -- Done"
+     echo "y"
+   else
+     echo "n"
+     log_info_msg "Cleanup mnesia and reset RMQ on $dst -- PENDING"
+   fi
+ fi
+}
+
+rmqclientsstop()
+{
+  for (( i=0; i<${DIPS_SIZE}; i++ ))
+   do
+    for (( j=0; j<${RMQ_CLNTS}; j++ ))
+     do
+      (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=5 ${DIPS[i]} "service ${RMQ_CLIENTS[j]} stop")
+      log_info_msg "${RMQ_CLIENTS[j]} stopped on ${DIPS[i]}"
+     done
+   done
+}
+
+rmqclientsrestart()
+{
+  for (( i=0; i<${DIPS_SIZE}; i++ ))
+   do
+    for (( j=0; j<${RMQ_CLNTS}; j++ ))
+     do
+      (ssh -o StrictHostKeyChecking=no -o ConnectTimeout=5 ${DIPS[i]} "service ${RMQ_CLIENTS[j]} restart")
+      log_info_msg "${RMQ_CLIENTS[j]} started on ${DIPS[i]}"
+     done
+   done
+}
+
+cleanpending() {
+if [ -f "$cleanuppending" ] && [ -s "$cleanuppending" ] && [[ $RABBITMQ_MNESIA_CLEAN == "True" ]]; then
+  readarray ips < $cleanuppending
+  ipsz=${#ips[@]}
+  for (( i=0; i<${ipsz}; i++ ))
+   do
+    status=$(ping -c 1 -w 1 -W 1 -n ${ips[i]} | grep packet | awk '{print $6}' | cut -c1)
+    if [[ $status == 0 ]]; then
+      rmqclientsstop
+      isClean=$(cleanup "${ips[i]}")
+      if [[ $isClean == "y" ]];  then
+        var=$(grep -F -ve "${ips[i]}" $cleanuppending)
+        echo $var > $cleanuppending
+        sed -i '/^$/d' $cleanuppending
+      fi
+    fi
+   done
+else
+  if [ -f "$cleanuppending" ] && [ ! -s "$cleanuppending" ]; then
+    pol=$($sethapolicy)
+    log_info_msg "HA Policy set - $pol"
+    rmqclientsrestart
+    (exec rm -rf "$cleanuppending")&
+  fi
+fi
 }
 
 checkNrst()
@@ -198,25 +288,47 @@ if [[ $chnlstate_run == "n" ]] || [[ $cluststate_run == "n" ]] || [[ $part_state
    if [[ $totalrst == '' ]]; then
        totalrst=0
    fi
-   if [ $totalrst == 2 ]; then
+   if [ $totalrst == 2 ] && [[ $RABBITMQ_MNESIA_CLEAN == "True" ]]; then
+    rmqclientsstop
+    cleanup "$MYIP"
     for (( i=0; i<${DIPS_SIZE}; i++ ))
      do
-       (ssh -o StrictHostKeyChecking=no ${DIPS[i]} "$rmqstop"; "$killbeam"; "$killepmd")&
-       (ssh -o StrictHostKeyChecking=no ${DIPS[i]} "$rmmnesia")&
-       (ssh -o StrictHostKeyChecking=no ${DIPS[i]} "$RMQ_RESET")&
-       log_info_msg "Cleaned up mnesia and reset all RMQ -- Done"
+       flag=true
+       if [ $MYIP != ${DIPS[i]} ]; then
+         status=$(ping -c 1 -w 1 -W 1 -n ${DIPS[i]} | grep packet | awk '{print $6}' | cut -c1)
+         if [[ $status == 0 ]]; then
+           isClean=$(cleanup "${DIPS[i]}")
+           if [[ $isClean == "n" ]]; then
+              flag=false
+           fi
+         else
+              flag=false
+         fi
+         if [ "$flag" == false ]; then
+           out=$(cat $cleanuppending | grep ${DIPS[i]})
+           if [[ $out == '' ]]; then
+             (exec echo ${DIPS[i]} >> "$cleanuppending")&
+           fi
+         fi
+       fi
      done
+     pol=$($sethapolicy)
+     log_info_msg "HA Policy set - $pol"
+     rmqclientsrestart
      (exec rm -rf "$numrst")&
    else
     if [ $cnt == 3 ]; then
      for (( i=0; i<${DIPS_SIZE}; i++ ))
      do
-       (ssh -o StrictHostKeyChecking=no ${DIPS[i]} "$RMQ_RESET")&
+       status=$(ping -c 1 -w 1 -W 1 -n ${DIPS[i]} | grep packet | awk '{print $6}' | cut -c1)
+       if [[ $status == 0 ]]; then
+           (ssh -o StrictHostKeyChecking=no ${DIPS[i]} "$RMQ_RESET")
+       fi
      done
      totalrst=$(($totalrst + 1))
      (exec echo $totalrst > "$numrst")&
+     log_info_msg "Tried resetting all available and connected RMQ -- Done"
      (exec rm -rf "$rstcnt")&
-     log_info_msg "Resetting all RMQ -- Done"
     else
      (exec $RMQ_RESET)&
      cnt=$(($cnt + 1))
@@ -230,8 +342,11 @@ fi
 (exec rm -rf "$cluschk")&
 (exec rm -rf "$cluspart")&
 (exec rm -rf "$rstinprog")&
+log_info_msg "check complete"
 }
 
+killstale()
+{
 stalels=$(ps -ef | grep rabbitmqctl | grep list_channels | awk '{print $2}')
 stalecs=$(ps -ef | grep rabbitmqctl | grep cluster_status | awk '{print $2}')
 if [[ $stalels != '' ]]; then
@@ -246,15 +361,49 @@ if [[ $stalecs != '' ]]; then
     (exec pkill -TERM -P $cpid)&
    done
 fi
+}
+
+function run_rmq_monitor()
+{
+ periodic_check
+ cleanpending
+ checkNrst
+ killstale
+}
+
+function run_onzk_lock_acquire {
+ZK_IPPORTS="$ZK_SERVER_IP" python - <<END
+import os
+import socket
+import subprocess
+import sys, getopt
+from kazoo.client import KazooClient
+zk_ip_ports=os.environ['ZK_IPPORTS']
+zk=KazooClient(zk_ip_ports)
+zk.start()
+lock=zk.Lock('/rmq-monitor','%s-%d' % (socket.gethostname(),os.getpid()))
+with lock:
+   subprocess.call("/opt/contrail/bin/contrail-rmq-monitor.sh MONITOR", shell=True)
+   lock.release()
+END
+}
 
 main()
 {
- lock $PROGNAME \
+ if [[ $command == $MONITOR ]]; then
+    run_rmq_monitor
+ else
+  lock $PROGNAME \
         || eexit "Only one instance of $PROGNAME can run at one time."
-
- periodic_check
- checkNrst
+  if [[ $command == $STOP ]]; then
+    $rmqstop
+    $killbeam
+    $killepmd
+  else
+    run_onzk_lock_acquire
+  fi
+ fi
+ exit 0
 }
-main
 
-exit 0
+main
